Support null values in database response
When executing a sql statement, each record of the result
is converted into a map. This map does not support null values,
instead it uses "NULL" strings, which cannot be handled the
right way by the caller.

To avoid this issue, the map now supports null values.
In this way the client is able to distinguish between strings
and columns without value.